feat: remove profile pictures and moved logout
